bsp: Switch to 4.4-cip kernel
Using the long-life kernel for the Civil Infrastructure Platform project
instead of the Yocto kernel comes with a number of advantages:

- long-term source for stable maintenance (10+ years)
- improved quality due to second review by cip kernel maintainer
- removal of license-wise problematic firmware files

This first version is based on current 4.4.98-cip plus two required
commits that are currently pending for integration. In addition, we just
have 5 local hacks that account for an issue in the BIOS and
dependencies of the current libmraa and the arduino sketches on the
original Intel BSP. Those will be eventually resolved.

We also switch to a defconfig based kernel configuration at this chance
which provides better control over our kernel. This configuration is
still subject to further tuning.

Currently, we keep our usage of meta-intel, now basically only for the
sake of obtaining its Quark processor tunings. This may be reconsidered
later on.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>